Remove ActiveRecord::Model from Oracle enhanced adapter
ActiveRecord::Model had been implemented in Rails master then reverted
in this commit.

Refer https://github.com/rails/rails/commit/9e4c41c903e8e58721f2c41776a8c60ddba7a0a9
